+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/board/dto/PostBoardDTO.java
@@ -1,4 +1,4 @@
-package backend.spectrum.dguonoff.domain.admin.dto;
+package backend.spectrum.dguonoff.domain.board.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
