+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/fixedSchedule/dto/DailyScheduleResponse.java
@@ -1,7 +1,7 @@
-package backend.spectrum.dguonoff.domain.admin.dto;
+package backend.spectrum.dguonoff.domain.fixedSchedule.dto;
 
-import backend.spectrum.dguonoff.domain.admin.dto.common.EventInfoDTO;
-import backend.spectrum.dguonoff.domain.admin.dto.common.PeriodDTO;
+import backend.spectrum.dguonoff.domain.fixedSchedule.dto.common.EventInfoDTO;
+import backend.spectrum.dguonoff.domain.fixedSchedule.dto.common.PeriodDTO;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
 import lombok.AllArgsConstructor;
