+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/fixedSchedule/repository/FixedScheduleRepository.java
@@ -1,4 +1,4 @@
-package backend.spectrum.dguonoff.domain.admin.repository;
+package backend.spectrum.dguonoff.domain.fixedSchedule.repository;
 
 
 import backend.spectrum.dguonoff.dao.FixedSchedule;
