+++ backend/dguonoff/src/main/java/backend/spectrum/dguonoff/domain/reservation/api/ReservationController.java
@@ -13,14 +13,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.mail.MessagingException;
 import java.security.Principal;
 import java.time.LocalDate;
 import java.util.List;
 
 import static backend.spectrum.dguonoff.global.statusCode.CommonCode.*;
-import static backend.spectrum.dguonoff.global.statusCode.CommonCode.SUCCESS_REJECTION;
-
 @RestController
 @Slf4j
 @RequestMapping("api/reservation")
@@ -28,7 +25,6 @@ import static backend.spectrum.dguonoff.global.statusCode.CommonCode.SUCCESS_REJ
 public class ReservationController {
 
     private final ReservationService reservationService;
-    private final UserService userService;
 
     //유저 자신의 예약 목록 조회 기능
     @GetMapping
@@ -39,32 +35,6 @@ public class ReservationController {
         HttpStatus successStatus = SUCCESS_FACILITY_LOOKUP.getStatus();
 
         return new ResponseEntity<>(reservationInfoList, successStatus);
-    }
-
-    //전체 예약 목록 조회
-    @GetMapping("/admin/all")
-    public ResponseEntity<List<ReservationInfoResponse>> getAllReservationInfo(Principal principal){
-        String adminId = principal.getName();
-
-        //관리자 권한 확인
-        userService.checkAdmin(adminId);
-
-        List<ReservationInfoResponse> reservationInfoList = reservationService.getAllReservationInfoList();
-
-        return new ResponseEntity<>(reservationInfoList, HttpStatus.OK);
-    }
-
-    // 특정 유저의 예약 목록 조회 기능
-    @GetMapping("admin/{userId}")
-    public ResponseEntity<List<ReservationInfoResponse>> getUserReservationInfo(@PathVariable String userId, Principal principal){
-        String adminId = principal.getName();
-
-        //관리자 권한 확인
-        userService.checkAdmin(adminId);
-
-        List<ReservationInfoResponse> reservationInfoList = reservationService.getReservationInfoList(userId);
-
-        return new ResponseEntity<>(reservationInfoList, HttpStatus.OK);
     }
 
     //예약 수정 기능
@@ -95,42 +65,6 @@ public class ReservationController {
 
         return new ResponseEntity<>(successMessage, successStatus);
     }
-
-
-    //관리자 예약 승인 기능
-    @GetMapping("/admin/approval/{reservationId}")
-    public ResponseEntity<String> approveReservation(@PathVariable Long reservationId, Principal principal) throws MessagingException {
-        String adminId = principal.getName();
-
-        //관리자 권한 확인
-        userService.checkAdmin(adminId);
-
-        //예약 승인
-        reservationService.approveReservation(reservationId);
-
-        String successMessage = String.format(SUCCESS_APPROVAL.getMessage(), adminId);
-        HttpStatus successStatus = SUCCESS_APPROVAL.getStatus();
-
-        return new ResponseEntity<>(successMessage, successStatus);
-    }
-
-    //관리자 예약 거절 기능
-    @GetMapping("/admin/reject/{reservationId}")
-    public ResponseEntity<String> rejectReservation(@PathVariable Long reservationId, Principal principal) throws MessagingException {
-        String adminId = principal.getName();
-
-        //관리자 권한 확인
-        userService.checkAdmin(adminId);
-
-        //예약 거절
-        reservationService.rejectReservation(reservationId);
-
-        String successMessage = String.format(SUCCESS_REJECTION.getMessage(), adminId);
-        HttpStatus successStatus = SUCCESS_REJECTION.getStatus();
-
-        return new ResponseEntity<>(successMessage, successStatus);
-    }
-
 
     //예약 가능 확인하기 기능
     @GetMapping("/available/{buildingName}/{facilityCode}/{date}")
